+++ packages/core/styles/src/gradients/gradients.ts
@@ -14,20 +14,25 @@
 import { varsThemeBase } from "../themes";
 import { createLinearGradient } from "./gradientBuilder";
 
-const interactiveVars = varsThemeBase.colors.aiGenerative.interactive;
+const { aiGradient } = varsThemeBase.colors;
 
 export const aiGenerativeInteractive = createLinearGradient(
-  "66deg",
-  `${interactiveVars.start} 28.22%`,
-  `${interactiveVars.middle} 49.27%`,
-  `${interactiveVars.end} 83.12%`
+  "50deg",
+  `${aiGradient.blueHigh} 28.22%`,
+  `${aiGradient.purpleHigh} 49.27%`,
+  `${aiGradient.pinkHigh} 83.12%`
 );
 
-const disabledVars = varsThemeBase.colors.aiGenerative.disabled;
+export const aiGenerativeHover = createLinearGradient(
+  "50deg",
+  `${aiGradient.pinkHigh} 28.22%`,
+  `${aiGradient.purpleHigh} 49.27%`,
+  `${aiGradient.blueHigh} 83.12%`
+);
 
 export const aiGenerativeDisabled = createLinearGradient(
-  "66deg",
-  `${disabledVars.start} 28.22%`,
-  `${disabledVars.middle} 49.27%`,
-  `${disabledVars.end} 83.12%`
+  "50deg",
+  `${aiGradient.blueLight} 28.22%`,
+  `${aiGradient.purpleLight} 49.27%`,
+  `${aiGradient.pinkLight} 83.12%`
 );