+++ packages/react/src/atomic/Button/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The Button component allows us to initiate actions, make state or page changes.
 
+## 2025-09-17 `2.8.0`
+
+#### 🎉 New features
+
+- New `ai-primary` and `ai-secondary` appearances to `Button` component to match Nimbus AI gradient style. ([#353](https://github.com/TiendaNube/nimbus-design-system/pull/353) by [@joacotornello](https://github.com/joacotornello))
+- Added `:hover` and `:active` properties to `Button` component AI State to improve hover and active states. ([#353](https://github.com/TiendaNube/nimbus-design-system/pull/353) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-08-04 `2.7.0`
 
 #### 🎉 New features