+++ packages/react/src/atomic/Textarea/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Textarea component allows the user to enter or edit information in long text format.
 
+## 2025-09-17 `2.4.1`
+
+#### 🎉 New features
+
+- Fixed `ai-generative` appearance to match Nimbus AI gradient style. ([#351](https://github.com/TiendaNube/nimbus-design-system/pull/351) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-09-15 `2.4.0`
 
 #### 🎉 New features